pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["Moncif"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.54.4"
streamlit = "^1.40.1"
python-dotenv = "^1.0.1"
trafilatura = "^1.12.2"
bs4 = "^0.0.2"
duckduckgo-search = "^6.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
